Create a UiUtils class and put dropdown_row in it
